Create goes right after
In a given word, you must check if a symbol appears only on the right after another.

Here are the cases you should expect when solving this challenge:

one of the symbols is not in the given word - your function must return False;
a symbol appears more than once in a word - use only the first one;
two symbols are identical - your function must return False;
the condition is case-sensitive, meaning that "a" and "A" are two different symbols.