Create max_letter_in_a_word
You are given a text containing various English letters and punctuation marks. You must find the most frequently occurring letter in the text. The letter found must be lowercase.
When searching for the most frequently occurring letter, case is not important, so for the purposes of your search, "A" = "a". Be careful not to count punctuation marks, numbers, and white spaces, only letters.