Create even_the_last
You are given a list of integers. You must find the sum of the elements with a pair of indices (0th, 2nd, 4th, etc.). Then multiply this summed number and the last element of the list together. Remember that the first element has an index of 0.

For a list, the result will always be 0 (zero).