Create median
A median is a numerical value that separates the upper half of a sorted list of numbers from the lower half. In a list where there are an odd number of entities, the median is the number in the middle of the list. If the list contains an even number of entities, there is no single middle value, but the median becomes the average of the two numbers in the middle. For this task, you are given a non-empty list of integers. With this array, you must separate the upper half of the numbers from the lower half and find the median.