Create follow_instructions
You've received a letter from a friend you haven't seen or heard from in a long time. In this letter, he gives you instructions to find a hidden treasure!

In this mission, you must follow a list of instructions that will take you to a specific point on the map. A list of instructions is a string; each letter in this string tells you the direction of your next step.

The letter "f" tells you to move forward, "b" tells you to move backward, "l" tells you to move left, and "r" tells you to move right.

Now imagine you are at position (0, 0). If you move forward, your position will become (0, 1). If you move forward again, your position will be (0, 2). If your next step is to the left, your position will become (-1, 2). After the next two steps, your coordinates will be (-1, 4).

Your goal is to find your final coordinates. In our case, it's (-1, 4).
This means that if your list of instructions is "fflff," you must move forward twice, take one step to the left, and move forward twice more.