Create remove_all_after
Not all elements are important. What you need to do here is delete all elements after the given element of list.

For example, let's say we have a list [1, 2, 3, 4, 5] and we need to delete all elements after 3—that is, 4 and 5.

We have two cases here:

If a cut element cannot be found, the list should not be modified;
If the list is empty, it should remain empty.
Input: List of integers and the border element.

Output: List or other Iterable (tuple, iterator, generator) of integers.